+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## \[Unreleased\]
 
+## Added
+
+- `Alert`:
+    - Added the following fields to `Alert.attributes`: `prv_sources`, `prv_sources.detections`, `ssobjectid`.
+
 ### Fixed
 
 - `Topic.publish()`: Message attributes were improperly defined